+++ server/controllers/hotel_controller.js
@@ -24,6 +24,10 @@ exports.hotel_details = (req, res) => {
 
 // Create a hotel
 exports.hotel_create = (req, res, next) => {
+  const { role } = req.user;
+  if (role !== 'admin')
+    return res.status(403).send({ status: 'error', error: 'Access denied' });
+
   const hotel = new Hotel(req.body);
   console.log(hotel);
   hotel.save((err) => {
@@ -37,6 +41,10 @@ exports.hotel_create = (req, res, next) => {
 
 // Update hotel
 exports.hotel_update = (req, res, next) => {
+  const { role } = req.user;
+  if (role !== 'admin')
+    return res.status(403).send({ status: 'error', error: 'Access denied' });
+
   Hotel.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -55,6 +63,12 @@ exports.hotel_update = (req, res, next) => {
 // Delete Hotel
 
 exports.hotel_delete = (req, res) => {
+  const { role } = req.user;
+  if (role !== 'admin')
+    return res
+      .status(403)
+      .send({ status: 'error', error: 'You must be logged in as admin user' });
+
   Hotel.findByIdAndRemove(req.params.id, (err) => {
     if (err) {
       res.send(err);
